Adds notes on CVE-2021-21571 and CVE-2021-21574
Adds notes on buffer overflow vulnerabilities, including an example of overflowing a buffer by providing a command line argument longer than the buffer size.

Adds notes on CVE-2021-21571, a vulnerability in the DELL Bios auto update feature and CVE-2021-21574, the vulnerability in the https connection that couses some other vulnerabilities. Explains how a crafted SSL certificate can be used to cause a buffer overflow in the BIOS.